Issue #52: finished and tested the test framework. Next step is to test on the ENA cluster for distributed test!

+++ roq-simulation/src/main/java/org/roqmessaging/loaders/TestLoaderDecription.java
@@ -159,7 +159,7 @@ public class TestLoaderDecription {
 		this.setDelay(
 				jsonObject.get("delay")!=null?((Long)jsonObject.get("delay")).intValue():5);
 		this.setDuration(
-				jsonObject.get("duration")!=null?((Long)jsonObject.get("duration")).intValue():1.0f);
+				jsonObject.get("duration")!=null?((Double)jsonObject.get("duration")).intValue():1.0f);
 		this.setMaxProd(
 				jsonObject.get("maxPub")!=null?((Long)jsonObject.get("maxPub")).intValue():1);
 		this.setMaxSub(
